% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{clean_sau_names}
\alias{clean_sau_names}
\title{Clean district/SAU names}
\usage{
clean_sau_names(names)
}
\arguments{
\item{names}{Character vector of names}
}
\value{
Cleaned character vector
}
\description{
Standardizes SAU (School Administrative Unit) names.
}
\keyword{internal}
