% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_raw_enrollment.R
\name{create_district_aggregate}
\alias{create_district_aggregate}
\title{Create district aggregate from school data}
\usage{
create_district_aggregate(school_df, end_year)
}
\arguments{
\item{school_df}{School-level data frame}

\item{end_year}{School year end}
}
\value{
District-level aggregated data frame
}
\description{
Create district aggregate from school data
}
\keyword{internal}
