% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_raw_enrollment.R
\name{create_state_from_districts}
\alias{create_state_from_districts}
\title{Create state row from district data}
\usage{
create_state_from_districts(df, end_year)
}
\arguments{
\item{df}{Processed district data}

\item{end_year}{School year end}
}
\value{
Single-row data frame with state totals
}
\description{
Create state row from district data
}
\keyword{internal}
