% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_raw_enrollment.R
\name{detect_year_from_data}
\alias{detect_year_from_data}
\title{Detect year from data}
\usage{
detect_year_from_data(df)
}
\arguments{
\item{df}{Data frame}
}
\value{
Integer year or NULL
}
\description{
Detect year from data
}
\keyword{internal}
