% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_raw_enrollment.R
\name{parse_maine_doe_excel}
\alias{parse_maine_doe_excel}
\title{Parse Maine DOE Excel file}
\usage{
parse_maine_doe_excel(file_path, end_year)
}
\arguments{
\item{file_path}{Path to downloaded Excel file}

\item{end_year}{School year end to extract}
}
\value{
List with school and district data frames
}
\description{
Extracts school and district data for a specific year from the
multi-year Maine DOE Excel file.
}
\keyword{internal}
