% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_raw_enrollment.R
\name{process_local_data}
\alias{process_local_data}
\title{Process local data to standard schema}
\usage{
process_local_data(df, end_year)
}
\arguments{
\item{df}{Raw data frame}

\item{end_year}{School year end}
}
\value{
Processed data frame
}
\description{
Process local data to standard schema
}
\keyword{internal}
