% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_raw_enrollment.R
\name{read_local_csv}
\alias{read_local_csv}
\title{Read local CSV file}
\usage{
read_local_csv(file_path, end_year)
}
\arguments{
\item{file_path}{Path to CSV file}

\item{end_year}{School year end}
}
\value{
Data frame
}
\description{
Read local CSV file
}
\keyword{internal}
