% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_raw_enrollment.R
\name{read_local_excel}
\alias{read_local_excel}
\title{Read local Excel file}
\usage{
read_local_excel(file_path, end_year)
}
\arguments{
\item{file_path}{Path to Excel file}

\item{end_year}{School year end}
}
\value{
Data frame
}
\description{
Read local Excel file
}
\keyword{internal}
