% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{standardize_grade}
\alias{standardize_grade}
\title{Standardize grade level names}
\usage{
standardize_grade(grade)
}
\arguments{
\item{grade}{Character vector of grade names}
}
\value{
Character vector of standardized grade names
}
\description{
Converts various grade level formats to standard form (PK, K, 01-12).
}
\keyword{internal}
