% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_raw_enrollment.R
\name{transform_maine_doe_data}
\alias{transform_maine_doe_data}
\title{Transform Maine DOE data to standard format}
\usage{
transform_maine_doe_data(df, end_year, level = "school")
}
\arguments{
\item{df}{Raw data frame from Excel}

\item{end_year}{Year to extract}

\item{level}{"school" or "district"}
}
\value{
Transformed data frame
}
\description{
Transform Maine DOE data to standard format
}
\keyword{internal}
