% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_raw_enrollment.R
\name{transform_state_level_data}
\alias{transform_state_level_data}
\title{Transform State Level data to standard format}
\usage{
transform_state_level_data(df, end_year)
}
\arguments{
\item{df}{Data frame from State Level sheet}

\item{end_year}{Year to extract}
}
\value{
Data frame with state-level enrollment
}
\description{
Transform State Level data to standard format
}
\keyword{internal}
