% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{validate_year}
\alias{validate_year}
\title{Validate year parameter}
\usage{
validate_year(end_year)
}
\arguments{
\item{end_year}{School year end}
}
\value{
TRUE if valid, otherwise throws an error
}
\description{
Checks if the provided year is within the available range.
}
\keyword{internal}
